smart_irrigation_system/node/requirements.txt
# Base (Node)
cbor2==5.6.5
certifi==2025.8.3
charset-normalizer==3.4.2
idna==3.10
luma.core==2.5.1
luma.oled==3.14.0
pillow==11.3.0
requests==2.32.4
smbus2==0.5.0
urllib3==2.5.0
rich==10.11.0
paho-mqtt>=1.6.1